CheapString prototype
